fix : 입력 숫자가 일의자리를 넘을 경우 숫자로 분리하는 로직 추가

+++ src/main/java/jumble_jump/domain/matcher/NumberMatcher.java
@@ -10,7 +10,7 @@ public class NumberMatcher implements Matcher{
         return new Number(validateNumber(input));
     }
 
-    private int validateNumber(char input) {
+    private double validateNumber(char input) {
         try {
             return Integer.parseInt(String.valueOf(input));
         } catch (NumberFormatException e) {
